refactors capability funcs
